Add files via upload
This code takes a string of DNA, chops them into triplet codons and scans for the initiator codon. If the initiator codon is found, the process of translation starts. The code takes DNA as reference, if needed, the Transcription code from my page can be taken and integrated into this program.